+++ CHANGELOG.md
@@ -4,8 +4,8 @@ Release notes for the Auto Translator Craft CMS plugin.
 
 ## Unreleased 1.0.0
 
-- Now requires PHP 8.2.0+.
-- Now requires Craft 5.0.0+.
+- Now compatible with PHP 8.2.0+.
+- Now compatible with Craft 5.0.0+.
 
 ## Unreleased 0.1.0